Move GetImageMoniker() Glyph extension method from Editor.wpf to EditorFeatures
