+++ src/VisualStudio/Core/Def/Implementation/FindReferences/Entries/Entry.cs
@@ -2,7 +2,6 @@
 
 using System.Windows;
 using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
-using Microsoft.CodeAnalysis.Editor.Wpf;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.VisualStudio.Shell.TableControl;
